requirements.txt
streamlit
utils
PyPDF2
docx2txt
python-dotenv
openai
pymupdf #fitz